Simplify ToolkitCredentialManager
* Remove interface since concrete implementation is only implementation
* Extract multiple class definitions into their own files
