+++ core/src/software/aws/toolkits/core/credentials/CredentialProviderNotFound.kt
@@ -1,0 +1,6 @@
+// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package software.aws.toolkits.core.credentials
+
+class CredentialProviderNotFound(msg: String) : RuntimeException(msg)
