+++ core/src/software/aws/toolkits/core/credentials/ToolkitCredentialsProviderManager.kt
@@ -8,33 +8,7 @@ import software.aws.toolkits.core.utils.tryOrNull
 import software.aws.toolkits.core.utils.warn
 import java.util.concurrent.ConcurrentHashMap
 
-interface ToolkitCredentialsChangeListener {
-    fun providerAdded(provider: ToolkitCredentialsProvider) {}
-    fun providerModified(provider: ToolkitCredentialsProvider) {}
-    fun providerRemoved(providerId: String) {}
-}
-
-interface ToolkitCredentialsProviderManager : ToolkitCredentialsChangeListener {
-    @Throws(CredentialProviderNotFound::class)
-    fun getCredentialProvider(id: String): ToolkitCredentialsProvider
-
-    /**
-     * Returns a list of all the registered providers. This list should never change for the life of the toolkit
-     */
-    fun getCredentialProviders(): List<ToolkitCredentialsProvider>
-
-    fun addChangeListener(listener: ToolkitCredentialsChangeListener)
-
-    fun removeChangeListener(listener: ToolkitCredentialsChangeListener)
-
-    /**
-     * Shuts down the manager and all registered factories
-     */
-    fun shutDown()
-}
-
-class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProviderRegistry) :
-    ToolkitCredentialsProviderManager {
+class ToolkitCredentialsProviderManager(registry: ToolkitCredentialsProviderRegistry) {
     private val listeners = ConcurrentHashMap.newKeySet<ToolkitCredentialsChangeListener>()
     private val factories = mutableListOf<ToolkitCredentialsProviderFactory<*>>()
 
@@ -43,20 +17,23 @@ class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProvi
     }
 
     @Throws(CredentialProviderNotFound::class)
-    override fun getCredentialProvider(id: String): ToolkitCredentialsProvider = factories.asSequence().mapNotNull { it.get(id) }.firstOrNull()
+    fun getCredentialProvider(id: String): ToolkitCredentialsProvider = factories.asSequence().mapNotNull { it.get(id) }.firstOrNull()
         ?: throw CredentialProviderNotFound("No ToolkitCredentialsProvider found represented by $id")
 
-    override fun getCredentialProviders(): List<ToolkitCredentialsProvider> = factories.flatMap { it.listCredentialProviders() }.toList()
+    /**
+     * Returns a list of all the registered providers. This list should never change for the life of the toolkit
+     */
+    fun getCredentialProviders(): List<ToolkitCredentialsProvider> = factories.flatMap { it.listCredentialProviders() }.toList()
 
-    override fun addChangeListener(listener: ToolkitCredentialsChangeListener) {
+    fun addChangeListener(listener: ToolkitCredentialsChangeListener) {
         listeners.add(listener)
     }
 
-    override fun removeChangeListener(listener: ToolkitCredentialsChangeListener) {
+    fun removeChangeListener(listener: ToolkitCredentialsChangeListener) {
         listeners.remove(listener)
     }
 
-    override fun providerAdded(provider: ToolkitCredentialsProvider) {
+    fun providerAdded(provider: ToolkitCredentialsProvider) {
         listeners.forEach {
             LOG.tryOrNull("Failed to notify listener that provider was added") {
                 it.providerAdded(provider)
@@ -64,7 +41,7 @@ class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProvi
         }
     }
 
-    override fun providerModified(provider: ToolkitCredentialsProvider) {
+    fun providerModified(provider: ToolkitCredentialsProvider) {
         listeners.forEach {
             LOG.tryOrNull("Failed to notify listener that provider was modified") {
                 it.providerModified(provider)
@@ -72,7 +49,7 @@ class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProvi
         }
     }
 
-    override fun providerRemoved(providerId: String) {
+    fun providerRemoved(providerId: String) {
         listeners.forEach {
             LOG.tryOrNull("Failed to notify listener that provider was removed") {
                 it.providerRemoved(providerId)
@@ -88,7 +65,7 @@ class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProvi
     /**
      * Shuts down the manager and all registered factories
      */
-    override fun shutDown() {
+    fun shutDown() {
         factories.forEach {
             try {
                 it.shutDown()
@@ -103,4 +80,3 @@ class DefaultToolkitCredentialsProviderManager(registry: ToolkitCredentialsProvi
     }
 }
 
-class CredentialProviderNotFound(msg: String) : RuntimeException(msg)
